refactor: move tools module to utils and update imports
* Renamed tools directory to utils for better naming convention
* Updated all imports from tools to utils across the codebase
* Moved json_helper, process, and time modules to utils directory
* Updated __init__.py in utils to include the moved modules
